Feat: updated Add0s.py
Changed the detection method from the entire name being to short to instead checking the episode number itself, this should allow shows with more than 10 seasons and entries with names like "series" or "season" instead.

Added some format checking to avoid non-episode files, checking that the file is indeed an mp4 and has a vaguely appropriate name and throwing if not.
The script then checks if the name is one of the aforementioned vaguely sensible names and reformats it if so, Note: since this changes the file name the file eg: "series1e1' can no longer be renamed so this reformat triggers a second pass to catch them.

Renamed to remove the space in the name